% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{addAllKeys}
\alias{addAllKeys}
\title{Add all keys from a file}
\usage{
addAllKeys(initData, keys, txtLines)
}
\arguments{
\item{initData}{The input data.frame.}

\item{keys}{A character vector with all the keys it is going to search for.}

\item{txtLines}{A character vector representation of the input file.}
}
\value{
A new data.frame with the new columns added.
}
\description{
This function adds all the keys found in a file to a given data.frame.
}

