% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{addKey}
\alias{addKey}
\title{Add values from key to data.frame}
\usage{
addKey(aDataFrame, aKey, refLines)
}
\arguments{
\item{aDataFrame}{The input data.frame.}

\item{aKey}{The key it is going to search for.}

\item{refLines}{A character vector representation of the input file.}
}
\value{
A new data.frame with the new column added.
}
\description{
This function adds a column to an existing data.frame. It names the column after the input key 
and then populates it with all the values it can find in the input file.
}
\keyword{internal}
