% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{addKeyFactory}
\alias{addKeyFactory}
\title{Add values closure}
\usage{
addKeyFactory(refLines)
}
\arguments{
\item{refLines}{A character vector representation of the input file.}
}
\value{
A function for adding keys to a data.frame.
}
\description{
This is a factory that returns a function that adds all key values to a data.frame, but for a
given and fixed character vector representing an input file.
}

