% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{class}
\name{AnnotationCommand-class}
\alias{AnnotationCommand-class}
\title{Base AnnotationCommand}
\description{
Abstract class that implements the execute interface for Command Design Pattern.
}
\section{Slots}{

\describe{
\item{\code{colName}}{Prefix used in order to generate the column name for the annotation.}
}}

