% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{class}
\name{AnnotationCommandList-class}
\alias{AnnotationCommandList-class}
\title{AnnotationCommandList}
\description{
This AnnotationCommand contains a list of several AnnotationCommands, and
executes them in order
}
\section{Slots}{

\describe{
\item{\code{commandList}}{A list containing AnnotationCommand objects.}
}}

