% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{annotationCommandList}
\alias{annotationCommandList}
\title{AnnotationCommandList constructor}
\usage{
annotationCommandList(...)
}
\arguments{
\item{...}{AnnotationCommand objects in the order to be processed.}
}
\description{
This function builds an AnnotationCommandList from a list of
AnnotationCommand commands
}

