% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{class}
\name{AtomicFilterCommand-class}
\alias{AtomicFilterCommand-class}
\title{AtomicFilterCommand}
\description{
Abstract class that serves as interface for all the filtering commands that are not lists. Allows
to define the filtering direction.
}
\section{Slots}{

\describe{
\item{\code{byRow}}{Indicates the direction of filtering, wheter rows or columns should be removed.}
}}

