% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\encoding{ISO-8859-2}
\name{averagePerBin}
\alias{averagePerBin}
\title{averagePerBin function from Hervé Pagès}
\usage{
averagePerBin(x, binsize, mcolnames)
}
\arguments{
\item{x}{A GenomicRanges object with non-NA seqlengths.}

\item{binsize}{A single positive integer.}

\item{mcolnames}{Names of numeric metadata columns in 'x' to "average" i.e. to propagate to the 
result after averaging them on each bin.}
}
\value{
A GRanges object with: (a) the same seqinfo as 'x', (b) ranges of width 'binsize' 
covering all the sequences in 'seqinfo(x)' and (c) the "averaged" metadata columns specified
in 'mcolnames'.
}
\description{
This function is an implementation of the code that Hervé Pagès generously provided at
http://permalink.gmane.org/gmane.science.biology.informatics.conductor/45929. It works by
averaging the specified metadata columns in 'mcolnames' using a given binsize over an input
ranges object.
}

