% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{barGroupGraph}
\alias{barGroupGraph}
\title{Bar Group Graph}
\usage{
barGroupGraph(dataList)
}
\arguments{
\item{dataList}{A list containing the bar groups. Each of the groups has to be of type character 
or factor. All the groups are then concatenated and converted to factors before generating the 
graph.}
}
\description{
This low-level function builds a basic bar graph from an input list. Each element of the list
corresponds to a bar in the graph, which is then segmented according to the factor defined by the
element.
}
