% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{barGroupGraphFromMatrix}
\alias{barGroupGraphFromMatrix}
\title{Bar Group Graph from matrix}
\usage{
barGroupGraphFromMatrix(dataMatrix)
}
\arguments{
\item{dataMatrix}{A matrix containing the bar group element counts. There is a colunn for each of
the bars/groups and a row for each level of the factor being studied.}
}
\description{
This low-level function builds a basic bar graph from an input matrix. The columns in the matrix
indicate the groups/bars, and the rows correspond to the different levels of the factor being 
studied. It is similar to the barGroupGraph function, but aimed to be used when the data list is
big.
}

