% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchplot.r
\name{batchPlot}
\alias{batchPlot}
\title{Batch effects plot}
\usage{
batchPlot(pdata, sv, nvars = NULL, qvars = NULL, alpha = 0.05)
}
\arguments{
\item{pdata}{A data.frame containing the phenotype information. Usually constructed by calling
the pData method on a MethylSet or GenomicMethylSet, or the colData method on a 
SummarizedExperiment.}

\item{sv}{A matrix containing the precomputed surogate variables.}

\item{nvars}{A character vector containing the names of the qualitative variables to analyze. If
NULL, its value is set to all the variables which happen to be of character or factor type.}

\item{qvars}{A character vector containing the names of the quantitative variables to analyze. If
NULL, its value is set to all the variables of numeric type. Be careful that this may not work as
expected, as some variables can be mistaken, such as the Slide number in a 450k experiment.}

\item{alpha}{Significance level for the qualitative relationship test.}
}
\value{
A data.frame containing the scores used to generate the plot.
}
\description{
A plot showing the association among both qualitative and quantitative variables and a set of
precomputed surogate variables. It is useful for evaluating the relationship among variables and 
the possible batch effects present in the data. For now, association between a quantitative 
variable and a surogate variable is measured by Pearson Correlation. For qualitative variables, 
an analysis of variance is performed and a tick is shown for the significant combinations.
}

