% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsizes.r
\name{camblorAC}
\alias{camblorAC}
\title{Camblor's AC}
\usage{
camblorAC(x, y, n = 512)
}
\arguments{
\item{x}{First sample.}

\item{y}{Second sample.}

\item{n}{Number of points to use for the densities estimation.}
}
\description{
This function computes Camblor's AC, a metric for evaluating the effect size in a non-parametric
way between two samples. It was implemented as a efficient replacement for the Cliff's delta in 
this package. Main difference is that AC is contained in the interval [0, 1] and it amounts to 
the area of the two densities (one for each of the input samples) intersection. So, a value of 1
would indicate that the two distributions are the same, hence there would be no effect size. A
value of 0 would indicate no intersection between distributions, and amount to a greater effect
size. It is an unsigned metric, so no information about the direction of the effect is provided.
}
