% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_tsts.r
\name{categoricalTest}
\alias{categoricalTest}
\title{General categorical testing.}
\usage{
categoricalTest(targetFactor, selectedIndices, testId = NULL)
}
\arguments{
\item{targetFactor}{A factor or character vector (which is then converted to a factor) encoding 
the universe for the test along with its labels.}

\item{selectedIndices}{A numeric, logical or character vector indicating the elements of the 
factor that resemble the subset currently being tested.}

\item{testId}{A character element identifying the current test.}
}
\value{
A list containing information about the result.
}
\description{
This function is a general building block for the common practice of testing if a given subset
of elements is differentially enriched or impovered with respect to a given factor. It builds
a contingency matrix, performs a statistical test and returns information about the test results
and effect sizes.
}
