% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epilab-datasets.r
\name{chenCohybTids}
\alias{chenCohybTids}
\title{List of sexual cohybridizing target IDs}
\format{A character vector containing 29233 target IDs}
\source{
\url{http://www.sickkids.ca/MS-Office-Files/Research/Weksberg Lab/48640-polymorphic-CpGs-Illumina450k.xlsx}
}
\description{
A character vector containing the Illumina450k target IDs of probes found
to cohybridate with sexual specific probes.
}
\references{
Chen YA, Lemire M, Choufani S, Butcher DT, Grafodatskaya D, 
  Zanke BW, Gallinger S, Hudson TJ, Weksberg R. Discovery of cross-reactive 
  probes and polymorphic CpGs in the Illumina Infinium HumanMethylation450 
  microarray. Epigenetics. 2013 Jan 11;8(2)
}

