% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsizes.r
\name{cliffDelta}
\alias{cliffDelta}
\title{Cliff's Delta}
\usage{
cliffDelta(x, y)
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample}
}
\description{
This function computes Cliff's Delta for two samples. It measures the effect size in a 
non parametric way. It is usually employed when a non-parametric test is employed, and a 
parametric effect size estimation, like Cohen's d, would make no sense. This naive implementation 
is quite computationally expensive, so it should be used with caution, as it could lead to memory
shortage problems.
}

