% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_tsts.r
\name{continuousTest}
\alias{continuousTest}
\title{General continuous testing.}
\usage{
continuousTest(targetVar, selectedIndices, testId = NULL)
}
\arguments{
\item{targetVar}{A numerical vector representing the values the universe variable takes.}

\item{selectedIndices}{A numeric, logical or character vector indicating the elements of the 
factor that resemble the subset currently being tested.}

\item{testId}{A character element identifying the current test.}
}
\value{
A list containing information about the result.
}
\description{
This function is a general building block for the common practice of testing if a given subset
of elements is differentially enriched or impovered with respect to a given numerical variable. 
It simply splits the numerical variable according to the indices provided, performs a statistical
test and returns information about the test results and effect sizes.
}
