% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{class}
\name{CPGICommand-class}
\alias{CPGICommand-class}
\title{CPGI Status AnnotationCommand}
\description{
This AnnotationCommand adds a column for the CpG Island status.
}
\section{Slots}{

\describe{
\item{\code{discardDirection}}{Logical value controlling whether the upstream(north)-downstream(south)
status of a CpG island-related region is important or not.}
}}

