% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{cpgiCommand}
\alias{cpgiCommand}
\title{CPGICommand constructor}
\usage{
cpgiCommand(colName, discardDirection = FALSE)
}
\arguments{
\item{colName}{Prefix used in order to generate the column name for the annotation.}

\item{discardDirection}{Logical value controlling whether the upstream(north)-downstream(south)
status of a CpG island-related region is important or not.}
}
\description{
This function builds a CPGICommand.
}

