% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{createFilesData}
\alias{createFilesData}
\title{Create initial data structure}
\usage{
createFilesData(txtLines)
}
\arguments{
\item{txtLines}{A character vector containing the lines of a text file.}
}
\value{
A data.frame with a single column containing all the identifiers.
}
\description{
This function reads the identifiers from a character vector representing a file and populates
an initial data.frame with their values.
}
\keyword{internal}
