% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{class}
\name{DensCpGCommand-class}
\alias{DensCpGCommand-class}
\title{Density of CpG AnnotationCommand}
\description{
This AnnotationCommand adds a column for the density of CpG around a given genomic region. The
size of the region is variable and controlled by the user.
}
\section{Slots}{

\describe{
\item{\code{windowSize}}{Size of the window (centered on the probe) for which the density is computed.}
}}

