% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{densCpGCommand}
\alias{densCpGCommand}
\title{DensCpG constructor}
\usage{
densCpGCommand(colName, windowSize = 2000)
}
\arguments{
\item{colName}{Prefix used in order to generate the column name for the annotation.}

\item{windowSize}{Size of the window centered on the input genomic region.}
}
\description{
This function builds a DensCpGCommand with a given windowSize.
}
