% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{class}
\name{DGenomicRegionCommand-class}
\alias{DGenomicRegionCommand-class}
\title{Disjoint Genomic Region AnnotationCommand}
\description{
This AnnotationCommand adds a column containing the genomic region for a given probe location. It
is different from the GenomicRegionCommand in that it generates disjoint categories and is more
specific.
}

