% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{dGenomicRegionCommand}
\alias{dGenomicRegionCommand}
\title{DGenomicRegionCommand constructor}
\usage{
dGenomicRegionCommand(colName)
}
\arguments{
\item{colName}{Prefix used in order to generate the column name for the annotation.}
}
\description{
This function builds a DGenomicRegionCommand with a given column name.
}

