% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epilab-package.r
\docType{package}
\name{epilab}
\alias{epilab}
\alias{epilab-package}
\title{EpiLab - Illumina Methylation Microrray}
\description{
Compilation of data, functions and other stuff intended to be used
inside the EpiLab for the implementation of data processing workflows. Most
of the functions are related to the analysis of Illumina 450k microarrays.
For now, the number of functions and methods is so low that I have decide
to keep everything inside only one package. In the future, this package
might be splitted to take into account the different aspects of the lab
workflows.
}
