% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{execute,AnnotationCommand,GRanges-method}
\alias{execute,AnnotationCommand,GRanges-method}
\title{AnnotationCommand base implementation of execute}
\usage{
\S4method{execute}{AnnotationCommand,GRanges}(command, object)
}
\arguments{
\item{command}{An AnnotationCommand object.}

\item{object}{A GRanges object containing the genomic regions to annotate.}
}
\description{
Implementation of the logic that is common to all AnnotationCommands. For
example, error management and input control. For now, it only prevents the command from being
executed on an empty object.
}

