% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{execute,AnnotationCommandList,GRanges-method}
\alias{execute,AnnotationCommandList,GRanges-method}
\title{AnnotationCommandList implementation of execute}
\usage{
\S4method{execute}{AnnotationCommandList,GRanges}(command, object)
}
\arguments{
\item{command}{An AnnotationCommandList.}

\item{object}{A GRanges object containing the genomic regions to annotate.}
}
\description{
An AnnotationCommandList is just a container for several AnnotationCommand objects. When
executed, the objects in the internal list are executed in the same order they were provided to
the constructor. The colName slot of AnnotationCommandList has to be empty, because it has no
useful meaning at the moment.
}
