% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{execute,CPGICommand,GRanges-method}
\alias{execute,CPGICommand,GRanges-method}
\title{CPGICommand implementation of execute}
\usage{
\S4method{execute}{CPGICommand,GRanges}(command, object)
}
\arguments{
\item{command}{A CPGICommand.}

\item{object}{A GRanges object containing the genomic regions to annotate.}
}
\description{
CPGICommand uses the CpG island definitions of Irizarry et al. in order to label the input
genomic regions according to their island status. Shores are defined as both 2kbp regions
flanking the island, and Shelves as the two external 2kbp regions flanking the Shores. As in the
DensCpGCommand, this command is best suited for the annotation of very short (methylation probes)
input regions.
}

