% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{execute,DensCpGCommand,GRanges-method}
\alias{execute,DensCpGCommand,GRanges-method}
\title{DensCpG implementation of execute}
\usage{
\S4method{execute}{DensCpGCommand,GRanges}(command, object)
}
\arguments{
\item{command}{A DensCpGCommand.}

\item{object}{A GRanges object containing the genomic regions to annotate.}
}
\description{
DensCpGCommand computes the density of CpG in a neighbourhood of the genomic input regions.
First, it resizes the input regions to the value of the windowSize slot. Be careful, however,
that in the case that the windowSize is smaller than the size of the input region the latter is
going to be narrowed by this action. For now, this is enough, because our input regions are
mostly GRanges objects containing methylation probes from Illumina microarrays. But it should be
fixed in order to use it for annotating NGS peaks, for example.
}
