% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{execute,DGenomicRegionCommand,GRanges-method}
\alias{execute,DGenomicRegionCommand,GRanges-method}
\title{DGenomicRegionCommand implementation of execute}
\usage{
\S4method{execute}{DGenomicRegionCommand,GRanges}(command, object)
}
\arguments{
\item{command}{A GenomicRegionCommand.}

\item{object}{A GRanges object containing the genomic regions to annotate.}
}
\description{
DGenomicRegionCommand tries to label the input genomic regions according to their relative
position with respect to the TSS. Regions defined include the Promoter region (2kbp upstream the
TSS), 5'UTR, First Exon, a different Exon, an Intron and the 3'UTR. Finally, Intergenic region is
assigned when neither of the former labels can be applied. This process is always executed at a
transcript level.
}
