% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{execute,FilterCommand,ANY-method}
\alias{execute,FilterCommand,ANY-method}
\title{FilterCommand base implementation of execute}
\usage{
\S4method{execute}{FilterCommand,ANY}(command, object)
}
\arguments{
\item{command}{A FilterCommand object.}

\item{object}{Any object for filtering.}
}
\description{
Implementation of the logic that is common to all FilterCommands. For 
example, error management and input control. For now, it only prevents the command from being
executed on an empty object.
}
