% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{execute,FilterCommandList,ANY-method}
\alias{execute,FilterCommandList,ANY-method}
\title{FilterCommandList implementation of execute}
\usage{
\S4method{execute}{FilterCommandList,ANY}(command, object)
}
\arguments{
\item{command}{A FilterCommandList.}

\item{object}{An object to be filtered.}
}
\description{
A FilterCommandList is just a container for several AtomicFilterCommand objects. When 
executed, the objects in the internal list are executed in the same order they were provided to
the constructor.
}

