% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{execute,GapCommand,GRanges-method}
\alias{execute,GapCommand,GRanges-method}
\title{GapCommand implementation of execute}
\usage{
\S4method{execute}{GapCommand,GRanges}(command, object)
}
\arguments{
\item{command}{A GapCommand.}

\item{object}{A GRanges object containing the genomic regions to annotate.}
}
\description{
GapCommand uses rtracklayer for connecting to the UCSC database and get the information
associated with the Gap track, which contains information about centromeres and telomeres.
Distance to the nearest telomere and centromere is then computed for the input regions.
}

