% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{execute,GenomicRegionCommand,GRanges-method}
\alias{execute,GenomicRegionCommand,GRanges-method}
\title{GenomicRegionCommand implementation of execute}
\usage{
\S4method{execute}{GenomicRegionCommand,GRanges}(command, object)
}
\arguments{
\item{command}{A GenomicRegionCommand.}

\item{object}{A GRanges object containing the genomic regions to annotate.}
}
\description{
GenomicRegionCommand tries to label the input genomic regions according to their relative
position with respect to the TSS. We define the Promoter region as the union of a 2kbp region
upstream the TSS, the first exon and the 5'UTR. Intragenic region is defined as the union of
the remaining exons and introns, those regions inside a gene which are not assigned to the
previous definition of Promoter region. Finally, Intergenic region is assigned when neither of
the former labels can be applied. This process is always executed at a transcript level, so
it is possible for an input region to be at Promoter and Intragenic regions at the same time.
Again, this command is better suited for very small input regions.
}
