% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{execute,KOverAFilterCommand,ANY-method}
\alias{execute,KOverAFilterCommand,ANY-method}
\title{KOverAFilterCommand implementation of execute for ANY}
\usage{
\S4method{execute}{KOverAFilterCommand,ANY}(command, object)
}
\arguments{
\item{command}{A KOverAFilterCommand command.}

\item{object}{An eSet object.}
}
\description{
KOverAFilterCommand discards rows or columns from the input object according to the 
internal detection p-values matrix of the command. It is equivalent to the kOverA function in the
genefilter package. A row/column is discarded if k or more elements have a detection p-value over
the parameter a. Direction of filtering is controlled by the byRow parameter.
}
\details{
In order to properly chain a list of commands (see FilterCommandList), the detection p-value
matrix is projected according to the row and column names of the object which is being filtered.
}
