% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{execute,MatrixFilterCommand,ANY-method}
\alias{execute,MatrixFilterCommand,ANY-method}
\title{MatrixFilterCommand implementation of execute for ANY}
\usage{
\S4method{execute}{MatrixFilterCommand,ANY}(command, object)
}
\arguments{
\item{command}{A MatrixFilterCommand command.}

\item{object}{An object to be filtered.}
}
\description{
This base implementation of execute just checks if the dimensions of the object to be filtered
are compatible with those from the additional matrix.
}
\details{
Dimensions do not have to be the same, but the row and column names of the input object should be 
included into row and column names of the additional matrix.
}

