% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{execute,NearestGeneCommand,GRanges-method}
\alias{execute,NearestGeneCommand,GRanges-method}
\title{NearestGeneCommand implementation of execute}
\usage{
\S4method{execute}{NearestGeneCommand,GRanges}(command, object)
}
\arguments{
\item{command}{A NearestGeneCommand.}

\item{object}{A GRanges object containing the genomic regions to annotate.}
}
\description{
NearestGeneCommand labels each input genomic region with information regarding to the nearest
gene. Gene information is obtained from the TxDb.Hsapiens.UCSC.hg19.knownGene transcripts
database. A gene region is defined as the union of all its transcript regions. The gene symbol
is obtained from the org.Hs.eg.db package.
}
