% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{execute,VarFilterCommand,ANY-method}
\alias{execute,VarFilterCommand,ANY-method}
\title{VarFilterCommand implementation of execute for ANY}
\usage{
\S4method{execute}{VarFilterCommand,ANY}(command, object)
}
\arguments{
\item{command}{A VarFilterCommand command.}

\item{object}{An object to be filtered.}
}
\description{
VarFilterCommand discards rows or columns according to their elements' variance. Depending on the
type of filtering, the threshold used can be absolute or expressed as a quantile of the whole
set of variances. This is a useful filter for the implementation of non-specific filtering.
}
