% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{execute}
\alias{execute}
\title{Execute generic}
\usage{
execute(command, object)
}
\arguments{
\item{command}{A command acting on an object and returning the same kind of object.}

\item{object}{The object the command acts on.}
}
\description{
Generic definition of the execute interface for a command on an object. An abstraction for
implementing hierarchies of commands based on the Command Design Pattern
}

