% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{class}
\name{FilterCommandList-class}
\alias{FilterCommandList-class}
\title{FilterCommandList}
\description{
This FilterCommand contains a list of several AtomicFilterCommands and executes them in order.
}
\section{Slots}{

\describe{
\item{\code{commandList}}{A list containing AtomicFilterCommand objects.}
}}

