% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{filterCommandList}
\alias{filterCommandList}
\title{FilterCommandList constructor.}
\usage{
filterCommandList(...)
}
\arguments{
\item{...}{AtomicFilterCommand objects in the order to be processed.}
}
\description{
This function builds a FilterCommandList from a list of AtomicFilterCommands.
}
