% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{first}
\alias{first}
\title{First closure}
\usage{
first(x)
}
\arguments{
\item{x}{An indexable object.}
}
\value{
The first element of x.
}
\description{
Just a closure for returning the first element of a vector.
}
\keyword{internal}
