% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{gapCommand}
\alias{gapCommand}
\title{GapCommand constructor}
\usage{
gapCommand(colName)
}
\arguments{
\item{colName}{Prefix used in order to generate the column name for the annotation.}
}
\description{
This function builds a GapCommand with a given column name.
}

