% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{generateCircosDensityFromRanges}
\alias{generateCircosDensityFromRanges}
\title{Generate CIRCOS density data from GRanges}
\usage{
generateCircosDensityFromRanges(ranges, ids = names(ranges), wsize = 1e+06,
  mcolname = NULL)
}
\arguments{
\item{ranges}{An input GRanges object. It must have sequence length information.}

\item{ids}{Character vector indicating the names of the elements to average over.}

\item{wsize}{Numeric value indicating the size of the window used for averaging.}

\item{mcolname}{Character vector indicating the name of the column for averaging.}
}
\value{
A data.frame with the averaged information in Circos friendly format.
}
\description{
This function generates a data.frame in CIRCOS compatible format from an input GRanges object.
In this case, an averaging step is performed in order to generate a windowed, averaged version
of a given input variable associated with the elements of the input genomic regions. It is also
possible to execute the function over a subset of element identifiers, and to specify one or
more metadata columns for averaging. It is useful for the display of genome-wide scale Circos
diagrams, when we want to see the information with a fixed zoom level.
}
