% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{generateCircosFromRanges}
\alias{generateCircosFromRanges}
\title{Generate CIRCOS suitable data from a GRanges (or a subset)}
\usage{
generateCircosFromRanges(ranges, ids = names(ranges), values = NULL)
}
\arguments{
\item{ranges}{The input GRanges object.}

\item{ids}{Character vector indicating the names of the elements to be generated.}

\item{values}{Numerical vector providing the corresponding elements' scores.}
}
\description{
This function is just a wrapper in order to generate data in the format that CIRCOS accepts. It 
allows to generate data from the whole GRanges object, or from a subset of it. It also accepts
a numerical vector of values to use as scores of the generated track information.
}

