% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{class}
\name{GenomicRegionCommand-class}
\alias{GenomicRegionCommand-class}
\title{Genomic Region AnnotationCommand}
\description{
This AnnotationCommand adds a column containing the genomic region for a given probe location.
}
