% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{genomicRegionCommand}
\alias{genomicRegionCommand}
\title{GenRegCommand constructor}
\usage{
genomicRegionCommand(colName)
}
\arguments{
\item{colName}{Prefix used in order to generate the column name for the annotation.}
}
\description{
This function builds a GenRegCommand with a given column name.
}

