% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{getA,KOverAFilterCommand-method}
\alias{getA,KOverAFilterCommand-method}
\title{KOverAFilterCommand base implementation of getA}
\usage{
\S4method{getA}{KOverAFilterCommand}(object)
}
\arguments{
\item{object}{A KOverAFilterCommand object.}
}
\description{
Getter method for a slot for all KOverAFilterCommand objects.
}

