% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{getAllKeys}
\alias{getAllKeys}
\title{Get all keys from file}
\usage{
getAllKeys(txtLines)
}
\arguments{
\item{txtLines}{A character vector containing the lines of a text file.}
}
\value{
A character vector containing all the keys found in the input.
}
\description{
This function retrieves all the different keys that are present in the input representation. As 
the number of keys and their names may be variable, it is necessary first to know which keys are 
present in the input in order to search for them afterwards.
}

