% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{getByRow,AtomicFilterCommand-method}
\alias{getByRow,AtomicFilterCommand-method}
\title{AtomicFilterCommand base implementation of getByRow}
\usage{
\S4method{getByRow}{AtomicFilterCommand}(object)
}
\arguments{
\item{object}{An AtomicFilterCommand object.}
}
\description{
Getter method for byRow slot for all AtomicFilterCommand objects.
}

