% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{getByRow}
\alias{getByRow}
\title{getByRow generic}
\usage{
getByRow(object)
}
\arguments{
\item{object}{An object that contains a byRow slot.}
}
\description{
Generic definition of the getByRow getter methods.
}

