% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{getColName,AnnotationCommand-method}
\alias{getColName,AnnotationCommand-method}
\title{AnnotationCommand base implementation of getColName}
\usage{
\S4method{getColName}{AnnotationCommand}(object)
}
\arguments{
\item{object}{An AnnotationCommand object.}
}
\description{
Getter method for colName slot for all AnnotationCommand objects.
}

