% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{getColName}
\alias{getColName}
\title{getColName generic}
\usage{
getColName(object)
}
\arguments{
\item{object}{An object that probably contains a colName slot.}
}
\description{
Generic definition of the getColName getter methods.
}

