% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{getCommandList,AnnotationCommandList-method}
\alias{getCommandList,AnnotationCommandList-method}
\title{AnnotationCommandList base implementation of getCommandList}
\usage{
\S4method{getCommandList}{AnnotationCommandList}(object)
}
\arguments{
\item{object}{An AnnotationCommandList object.}
}
\description{
Getter method for commandList slot for all AnnotationCommandList objects.
}

