% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{getCommandList,FilterCommandList-method}
\alias{getCommandList,FilterCommandList-method}
\title{FilterCommandList base implementation of getCommandList generic.}
\usage{
\S4method{getCommandList}{FilterCommandList}(object)
}
\arguments{
\item{object}{A FilterCommandList object.}
}
\description{
Getter method for commandList slot for all FilterCommandList objects.
}
