% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{getCommandList}
\alias{getCommandList}
\title{getCommandList generic}
\usage{
getCommandList(object)
}
\arguments{
\item{object}{An object that probably contains a commandList slot.}
}
\description{
Generic definition of the getCommandList getter methods.
}
