% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{getDiscardDirection,CPGICommand-method}
\alias{getDiscardDirection,CPGICommand-method}
\title{CPGICommand base implementation of getDiscardDirection}
\usage{
\S4method{getDiscardDirection}{CPGICommand}(object)
}
\arguments{
\item{object}{An CPGICommand object.}
}
\description{
Getter method for discardDirection slot for all CPGICommand objects.
}

