% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{getDiscardDirection}
\alias{getDiscardDirection}
\title{getDiscardDirection generic}
\usage{
getDiscardDirection(object)
}
\arguments{
\item{object}{An object that probably contains a discardDirection slot.}
}
\description{
Generic definition of the getDiscardDirection getter methods.
}

