% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probegene.r
\name{getGeneEntrezIds}
\alias{getGeneEntrezIds}
\title{Get gene identifiers for a given set of target Ids}
\usage{
getGeneEntrezIds(tids, ...)
}
\arguments{
\item{tids}{A character vector containing the input Target IDs.}

\item{...}{Parameters to be passed to getProbeGeneRelationship.}
}
\value{
A character vector containing the gene identifier.
}
\description{
This function takes as input a given set of Illumina 450k target IDs and produces a list of gene
Entrez IDs, according to some parameters governing the getProbeGeneRelationship inner function.
}
