% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{getK,KOverAFilterCommand-method}
\alias{getK,KOverAFilterCommand-method}
\title{KOverAFilterCommand base implementation of getK}
\usage{
\S4method{getK}{KOverAFilterCommand}(object)
}
\arguments{
\item{object}{An KOverAFilterCommand object.}
}
\description{
Getter method for k slot for all KOverAFilterCommand objects.
}

