% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{getMatrix,MatrixFilterCommand-method}
\alias{getMatrix,MatrixFilterCommand-method}
\title{MatrixFilterCommand base implementation of getMatrix}
\usage{
\S4method{getMatrix}{MatrixFilterCommand}(object)
}
\arguments{
\item{object}{A MatrixFilterCommand object.}
}
\description{
Getter method for m slot for all MatrixFilterCommand objects.
}

