% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{getMatrix}
\alias{getMatrix}
\title{getMatrix generic}
\usage{
getMatrix(object)
}
\arguments{
\item{object}{An object that contains a m slot.}
}
\description{
Generic definition of the getMatrix getter methods.
}
