% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probegene.r
\name{getProbeGeneRelationship}
\alias{getProbeGeneRelationship}
\title{Get relationship between probes and genes}
\usage{
getProbeGeneRelationship(tids, method = c("illumina", "ucsc19"),
  promoterSize = 2000)
}
\arguments{
\item{tids}{A character vector containing the input target ID's.}

\item{method}{Gene overlap computing method.}

\item{promoterSize}{Size to add to transcripts' upstream promoter regions.}
}
\value{
A data.frame containing four columns: targetId, txId, geneId and geneSymbol.
}
\description{
This function computes the overlapping between Illumina450k probes and a set of genes determined
by a given method. The 'illumina' method just uses the annotation information in the
IlluminaHumanMethylation450k.db package, while the 'ucsc19' method uses the information from
UCSC hg19 genome version and computes the overlaps at a transcript level. There is also the
possibility to define a given promoter size to also compute the overlapping with these special
regions. The function uses the getSymbolsFromEntrezIds function from this package in order to
get the genes' names. This function returns a data.frame with all the overlaps corresponding
to the input target Id's.
}

