% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probegene.r
\name{getSymbolsFromEntrezIds}
\alias{getSymbolsFromEntrezIds}
\title{Get gene symbols for a given set of Entrez IDs}
\usage{
getSymbolsFromEntrezIds(entrezIds)
}
\arguments{
\item{entrezIds}{A character vector containing the Entrez IDs.}
}
\description{
This function takes as input a set of Entrez IDs and generates a list of gene symbols using the
annotation information from the org.Hs.eg.db Bioconductor package.
}

