% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{getThreshold,VarFilterCommand-method}
\alias{getThreshold,VarFilterCommand-method}
\title{VarFilterCommand base implementation of getThreshold}
\usage{
\S4method{getThreshold}{VarFilterCommand}(object)
}
\arguments{
\item{object}{An VarFilterCommand object.}
}
\description{
Getter method for threshold slot for all VarFilterCommand objects.
}
