% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{getThreshold}
\alias{getThreshold}
\title{getThreshold generic}
\usage{
getThreshold(object)
}
\arguments{
\item{object}{An object that contains a threshold slot.}
}
\description{
Generic definition of the getThreshold getter methods.
}

