% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{getType,VarFilterCommand-method}
\alias{getType,VarFilterCommand-method}
\title{VarFilterCommand base implementation of getType}
\usage{
\S4method{getType}{VarFilterCommand}(object)
}
\arguments{
\item{object}{An VarFilterCommand object.}
}
\description{
Getter method for type slot for all VarFilterCommand objects.
}

