% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{getType}
\alias{getType}
\title{getType generic}
\usage{
getType(object)
}
\arguments{
\item{object}{An object that contains a type slot.}
}
\description{
Generic definition of the getType getter methods.
}
