% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{getValuesForKey}
\alias{getValuesForKey}
\title{Get all values for a key}
\usage{
getValuesForKey(key, txts)
}
\arguments{
\item{key}{A character value representing the value we are looking for.}

\item{txts}{A character vector containing the lines of a text file.}
}
\value{
A character vector with all the values corresponding to the input key.
}
\description{
This function retrieves all values associated to a given key from a character vector representing
a text file. This text file is encoded as an identifier and a variable set of keys per line. The
keys are in the form key=value; while the identifier is the first element in the line and it is 
separated from the rest of elements with a tab character.
}
