% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{getWindowSize,DensCpGCommand-method}
\alias{getWindowSize,DensCpGCommand-method}
\title{DensCpGCommand base implementation of getWindowSize}
\usage{
\S4method{getWindowSize}{DensCpGCommand}(object)
}
\arguments{
\item{object}{An DensCpGCommand object.}
}
\description{
Getter method for windowSize slot for all AnnotationCommand objects.
}
