% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{getWindowSize}
\alias{getWindowSize}
\title{getWindowSize generic}
\usage{
getWindowSize(object)
}
\arguments{
\item{object}{An object that probably contains a windowSize slot.}
}
\description{
Generic definition of the getWindowSize getter methods.
}
