% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{class}
\name{KOverAFilterCommand-class}
\alias{KOverAFilterCommand-class}
\title{KOverAFilterCommand filter command.}
\description{
This FilterCommand filters out the rows/columns of the input object when there are at least K
elements over a value A in a given row/column. It is equivalent to the kOverA filter function in
the package genefilter.
}
