% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{kOverAFilterCommand}
\alias{kOverAFilterCommand}
\title{KOverAFilterCommand constructor}
\usage{
kOverAFilterCommand(m, byRow = TRUE, k = ifelse(byRow, 5, 5000), a = 0.01)
}
\arguments{
\item{m}{A matrix used for filtering.}

\item{byRow}{A logical indicating the direction of filtering.}

\item{k}{The minimum number of elements to label a row/column as invalid.}

\item{a}{The p-value threshold.}
}
\description{
This function builds a KOverAFilterCommand from parameters k and a.
}
