% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{kOverAFilterCommandFromFraction}
\alias{kOverAFilterCommandFromFraction}
\title{KOverAFilterCommand percentage constructor}
\usage{
kOverAFilterCommandFromFraction(m, byRow = TRUE, fraction = 0.1, a = 0.01)
}
\arguments{
\item{m}{A matrix used for filtering.}

\item{byRow}{A logical indicating the direction of filtering.}

\item{fraction}{The minimum percentage of elements to label a row/column as invalid.}

\item{a}{The p-value threshold.}
}
\description{
This function builds a KOverAFilterCommand from parameter a and a percentage of rows/columns.
}
