% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{matchLength}
\alias{matchLength}
\title{matchLength closure}
\usage{
matchLength(x)
}
\arguments{
\item{x}{An object with a match.length attribute.}
}
\value{
The value of the match.length attribute.
}
\description{
Just a closure for returning the match.length attribute of an object.
}
