% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{class}
\name{MatrixFilterCommand-class}
\alias{MatrixFilterCommand-class}
\title{MatrixFilterCommand}
\description{
Abstract class that serves as interface for all the filtering commands that use information from
an additional matrix. In our case, this is usually a matrix of detection p-values, and it helps
us to clean a set according to criteria defined over the input objec and the detection p-values
matrix.
}
\section{Slots}{

\describe{
\item{\code{m}}{The additional matrix used in the filtering criterium.}
}}

