% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{class}
\name{NearestGeneCommand-class}
\alias{NearestGeneCommand-class}
\title{Nearest gene AnnotationCommand}
\description{
This AnnotationCommand adds columns with information regarding the nearest TSS, TX and gene.
}

