% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{nearestTSSGeneCommand}
\alias{nearestTSSGeneCommand}
\title{nearestTSSGeneCommand constructor}
\usage{
nearestTSSGeneCommand(colName)
}
\arguments{
\item{colName}{Prefix used in order to generate the column name for the annotation.}
}
\description{
This function builds a NearestGeneCommand with a given column name and the type of nearest (TSS)
}
