% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsizes.r
\name{oddsRatio}
\alias{oddsRatio}
\title{Odds Ratio}
\usage{
oddsRatio(x, changeRows = FALSE)
}
\arguments{
\item{x}{Input 2x2 matrix.}

\item{changeRows}{boolean indicating if rows have to be exchanged prior to OR computation.}
}
\description{
This function computes the odds ratio for a 2x2 matrix. It is assumed that the true positive 
element count is located in the main diagonal. If that is not the case, the additional parameter
changeRows can be used to reflect that situation.
}

