% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsizes.r
\name{oddsRatioLevel}
\alias{oddsRatioLevel}
\title{Odds Ratio for a given level}
\usage{
oddsRatioLevel(bigTable, aLevel)
}
\arguments{
\item{bigTable}{The nx2 input matrix.}

\item{aLevel}{The level to be studied. It must be a valid rowname for the matrix.}
}
\description{
It is not uncommon to work with nx2 contingency matrices, and it is quite often interesting to 
be able to compute the odds ratios at the different levels the discrete variable can take. This
function does exactly that. It just collapses a bigger matrix onto a 2x2 matrix, like a binary
variable representing the pertenence of the data to the given level. The it computes the odds
ratio in the traditional way.
}
