% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_tsts.r
\name{rangeTestAgainstList}
\alias{rangeTestAgainstList}
\title{Test ranges against background with respect to a list of ranges.}
\usage{
rangeTestAgainstList(selectedRange, backgroundRange, rangeList, ...)
}
\arguments{
\item{selectedRange}{A GRanges object containing the sequences of interest.}

\item{backgroundRange}{A GRanges object containing sequences which are not supposed to resemble
the selected subset.}

\item{rangeList}{A GRangesList object containing the different subsets we are going to test 
against.}

\item{\dots}{Parameters to be passed down to mclapply.}
}
\value{
A data.frame containing the results from all the tests.
}
\description{
This function traverses the provided GRangesList and, for each element of the list, it performs
a simple proportionality test based on a 2x2 contingency table. This is similar to a gene set
testing (GO-like), but a lower level and intended for being used as a building block for more
complex testing procedures. Currently, Fisher test is the only choice.
}

