% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{readBroadHistoneFile}
\alias{readBroadHistoneFile}
\title{Read a variable key, value file}
\usage{
readBroadHistoneFile(txtLines)
}
\arguments{
\item{txtLines}{A character vector representation of the input file.}
}
\value{
A new data.frame with the new columns added.
}
\description{
The files describing the ENCODE Broad Histone experiments are in a special format, where for each
line, there is an identifier followed by a variable set of key, value pairs encoding the 
necessary information. This function reads the information in a file and returns a data.frame
with a column for each key.
}
