% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{second}
\alias{second}
\title{Second closure}
\usage{
second(x)
}
\arguments{
\item{x}{An indexable object.}
}
\value{
The second element of x.
}
\description{
Just a closure for returning the second element of a vector.
}
\keyword{internal}
