% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{secondLength}
\alias{secondLength}
\title{secondLength closure}
\usage{
secondLength(x)
}
\arguments{
\item{x}{An object with a match.length attribute.}
}
\value{
The value of the second element of the match.length attribute.
}
\description{
Just a closure for returning the second element of a match.length attribute of an object.
}
\keyword{internal}
