% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{setA<-,KOverAFilterCommand-method}
\alias{setA<-,KOverAFilterCommand-method}
\title{KOverAFilterCommand base implementation of setA}
\usage{
\S4method{setA}{KOverAFilterCommand}(object) <- value
}
\arguments{
\item{object}{A KOverAFilterCommand object.}

\item{value}{The new a value.}
}
\description{
Getter method for a slot for all KOverAFilterCommand objects.
}

