% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{setA<-}
\alias{setA<-}
\title{setA generic}
\usage{
setA(object) <- value
}
\arguments{
\item{object}{An object that contains a a slot.}

\item{value}{The new a value.}
}
\description{
Generic definition of the setA setter methods.
}
