% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{setByRow<-,AtomicFilterCommand-method}
\alias{setByRow<-,AtomicFilterCommand-method}
\title{AtomicFilterCommand base implementation of setByRow}
\usage{
\S4method{setByRow}{AtomicFilterCommand}(object) <- value
}
\arguments{
\item{object}{An AtomicFilterCommand object.}

\item{value}{The new byRow value.}
}
\description{
Getter method for byRow slot for all AtomicFilterCommand objects.
}
