% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{setByRow<-}
\alias{setByRow<-}
\title{setByRow generic}
\usage{
setByRow(object) <- value
}
\arguments{
\item{object}{An object that contains a byRow slot.}

\item{value}{The new byRow value.}
}
\description{
Generic definition of the setByRow setter methods.
}

