% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\docType{methods}
\name{setColName<-,AnnotationCommand-method}
\alias{setColName<-,AnnotationCommand-method}
\title{AnnotationCommand base implementation of setColName}
\usage{
\S4method{setColName}{AnnotationCommand}(object) <- value
}
\arguments{
\item{object}{An AnnotationCommand object.}

\item{value}{The new column name.}
}
\description{
Getter method for colName slot for all AnnotationCommand objects.
}
