% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{setColName<-}
\alias{setColName<-}
\title{setColName generic}
\usage{
setColName(object) <- value
}
\arguments{
\item{object}{An object that probably contains a colName slot.}

\item{value}{The new column name.}
}
\description{
Generic definition of the setColName setter methods.
}
