% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{setK<-,KOverAFilterCommand-method}
\alias{setK<-,KOverAFilterCommand-method}
\title{KOverAFilterCommand base implementation of setK}
\usage{
\S4method{setK}{KOverAFilterCommand}(object) <- value
}
\arguments{
\item{object}{An KOverAFilterCommand object.}

\item{value}{The new k value.}
}
\description{
Getter method for k slot for all KOverAFilterCommand objects.
}

