% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{setK<-}
\alias{setK<-}
\title{setK generic}
\usage{
setK(object) <- value
}
\arguments{
\item{object}{An object that contains a k slot.}

\item{value}{The new k value.}
}
\description{
Generic definition of the setK setter methods.
}

