% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{setMatrix<-,MatrixFilterCommand-method}
\alias{setMatrix<-,MatrixFilterCommand-method}
\title{MatrixFilterCommand base implementation of setMatrix}
\usage{
\S4method{setMatrix}{MatrixFilterCommand}(object) <- value
}
\arguments{
\item{object}{A MatrixFilterCommand object.}

\item{value}{The new m value.}
}
\description{
Getter method for m slot for all MatrixFilterCommand objects.
}
