% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{setMatrix<-}
\alias{setMatrix<-}
\title{setMatrix generic}
\usage{
setMatrix(object) <- value
}
\arguments{
\item{object}{An object that contains a m slot.}

\item{value}{The new m value.}
}
\description{
Generic definition of the setMatrix setter methods.
}

