% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{setThreshold<-,VarFilterCommand-method}
\alias{setThreshold<-,VarFilterCommand-method}
\title{VarFilterCommand base implementation of setThreshold}
\usage{
\S4method{setThreshold}{VarFilterCommand}(object) <- value
}
\arguments{
\item{object}{An VarFilterCommand object.}

\item{value}{The new threshold value.}
}
\description{
Getter method for threshold slot for all VarFilterCommand objects.
}

