% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{setThreshold<-}
\alias{setThreshold<-}
\title{setThreshold generic}
\usage{
setThreshold(object) <- value
}
\arguments{
\item{object}{An object that contains a threshold slot.}

\item{value}{The new threshold value.}
}
\description{
Generic definition of the setThreshold setter methods.
}

