% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{methods}
\name{setType<-,VarFilterCommand-method}
\alias{setType<-,VarFilterCommand-method}
\title{VarFilterCommand base implementation of setType}
\usage{
\S4method{setType}{VarFilterCommand}(object) <- value
}
\arguments{
\item{object}{An VarFilterCommand object.}

\item{value}{The new type value.}
}
\description{
Getter method for type slot for all VarFilterCommand objects.
}

