% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{setType<-}
\alias{setType<-}
\title{setType generic}
\usage{
setType(object) <- value
}
\arguments{
\item{object}{An object that contains a type slot.}

\item{value}{The new type value.}
}
\description{
Generic definition of the setType setter methods.
}

