% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{simpleBarGraph}
\alias{simpleBarGraph}
\title{Simple Bar Graph}
\usage{
simpleBarGraph(dataVector)
}
\arguments{
\item{dataVector}{A named vector containing the counts for each group.}
}
\description{
This function builds a simple count bar plot from a named numeric vector. Each element of the
vector represents the number of elements in a group (bar). The group is named after the name of
the vector element.
}
