% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{class}
\name{SimpleFilterCommand-class}
\alias{SimpleFilterCommand-class}
\title{SimpleFilterCommand}
\description{
Abstract class that serves as interface for all the simple filtering commands that only use 
information from the object to be filtered.
}

