% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_tsts.r
\name{tids450kTestAgainstList}
\alias{tids450kTestAgainstList}
\title{Test Illumina450k target ids against a list of ranges.}
\usage{
tids450kTestAgainstList(targetIds, rangeList, ...)
}
\arguments{
\item{targetIds}{A character vector containing the Illumina450k target identifiers.}

\item{rangeList}{A GRangesList object containing the different subsets we are going to test 
against.}

\item{\dots}{Parameters to be passed down to rangeTestAgainstList.}
}
\value{
A data.frame containing the results from all the tests.
}
\description{
This function is a wrapper for the rangeTestAgainstList, specifically designed to handle
Illumina450k Target IDs. It internally transforms the ids into both selected and background
GRanges objects and then call the main workhorse function.
}

