% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{updateSeqLengthsFromBSGenome}
\alias{updateSeqLengthsFromBSGenome}
\title{Fill sequence length information}
\usage{
updateSeqLengthsFromBSGenome(ranges, type = c("hg19"))
}
\arguments{
\item{ranges}{A GRanges input object.}

\item{type}{Character value indicating the annotation to use.}
}
\value{
The same GRanges with the seqlengths field updated for the given annotation.
}
\description{
This function fills the seqlength field of a given GRanges object with information extracted 
from several sources, mainly a BSgenome annotation package. This is just a convenience function
for some of our workflows, and it currently supports only hg19 annotation package.
}

