% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\docType{class}
\name{VarFilterCommand-class}
\alias{VarFilterCommand-class}
\title{VarFilterCommand filter command}
\description{
This FilterCommand filters out the rows (or columns) of the input object where the variance meets
a certain requirement. Two criteria are implemented at the moment: filter the rows (or columns)
where the variance is under a given threshold, or filter the rows (or columns) where the variance
is under a given quantile of all elements' variances.
}
\section{Slots}{

\describe{
\item{\code{type}}{Indicates type of filtering. Should be one of 'quantile' or 'absolute'}

\item{\code{threshold}}{The numerical value of the threshold.}
}}

