% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.r
\name{varFilterCommand}
\alias{varFilterCommand}
\title{VarFilterCommand constructor}
\usage{
varFilterCommand(m, byRow = TRUE, type = "quantile", threshold = 0.25)
}
\arguments{
\item{m}{An additional matrix for evaluating the variance.}

\item{byRow}{A logical indicating the direction of filtering.}

\item{type}{Indicates type of filtering. Should be one of 'quantile' or 'absolute'}

\item{threshold}{The numerical value of the threshold.}
}
\description{
This function builds a VarFilterCommand from parameters type and threshold.
}
