% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.r
\name{violinGroupGraph}
\alias{violinGroupGraph}
\title{Violin Group Graph}
\usage{
violinGroupGraph(dataList)
}
\arguments{
\item{dataList}{A list containing the violin groups. Each of the groups has to be of type 
numeric.}
}
\description{
This function builds a basic violin plot from a data list. Each element of the list will be
represented by a different violin and named after its list name. The elements of the list have to
be numerical vectors.
}

