% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparse.r
\name{whichOne}
\alias{whichOne}
\title{WhichOne functional}
\usage{
whichOne(x, i)
}
\arguments{
\item{x}{An indexable object.}

\item{i}{Integer value indicating a position.}
}
\value{
The i'th element of x.
}
\description{
This function just returns a given element of a vector.
}

