% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos.r
\name{writeCircos}
\alias{writeCircos}
\title{Write CIRCOS data}
\usage{
writeCircos(circosData, filename)
}
\arguments{
\item{circosData}{A data.frame containing information in CIRCOS format.}

\item{filename}{Name of the file to write the information to.}
}
\description{
This is just a wrapper around write.table to ensure that CIRCOS data gets written to disk in good
shape. Just a way of grouping the formatting options and isolating the side effects for testing.
}
